Moving annotation to archiver
