+++ src/rachis/core/tests/test_cache.py
@@ -691,25 +691,25 @@ class TestCache(unittest.TestCase):
 
         with cache2:
             art1 = rachis.sdk.Result.load(art1_path)
-            self.assertEqual(art1._annotations, {})
+            self.assertEqual(art1._archiver._annotations, {})
             Note1 = rachis.core.annotate.Note(name='annotation-1', text='1')
             art1.add_annotation(Note1)
             self.cache.save(art1, 'annotation-1')
             art1_annotation1 = self.cache.load('annotation-1')
 
         self.assertEqual(
-            set(art1_annotation1._annotations.keys()), set(['annotation-1']))
+            set(art1_annotation1._archiver._annotations.keys()), set(['annotation-1']))
 
         with cache3:
             art1_copy = rachis.sdk.Result.load(art1_copy_path)
-            self.assertEqual(art1_copy._annotations, {})
+            self.assertEqual(art1_copy._archiver._annotations, {})
             Note2 = rachis.core.annotate.Note(name='annotation-2', text='2')
             art1_copy.add_annotation(Note2)
             self.cache.save(art1_copy, 'annotation-2')
             art1_both_annotations = self.cache.load('annotation-2')
 
         self.assertEqual(
-            set(art1_both_annotations._annotations.keys()),
+            set(art1_both_annotations._archiver._annotations.keys()),
             set(['annotation-1', 'annotation-2']))
 
     def test_annotation_merge_collisions(self):
@@ -732,7 +732,9 @@ class TestCache(unittest.TestCase):
             art1_annotation1 = self.cache.load('annotation-1')
 
         self.assertEqual(
-            set(art1_annotation1._annotations.keys()), set(['annotation-1']))
+            set(art1_annotation1._archiver._annotations.keys()),
+            set(['annotation-1'])
+        )
 
         with cache3:
             art1_copy = rachis.sdk.Result.load(art1_copy_path)
@@ -744,5 +746,5 @@ class TestCache(unittest.TestCase):
             art1_both_annotations = self.cache.load('annotation-1')
 
         self.assertEqual(
-            set(art1_both_annotations._annotations.keys()),
+            set(art1_both_annotations._archiver._annotations.keys()),
             set(['annotation-1', f'annotation-1-{Note2.id}']))
