+++ src/rachis/sdk/result.py
@@ -159,29 +159,6 @@ class Result(IResult):
     def citations(self):
         return self._archiver.citations
 
-    @property
-    def _annotations(self):
-        """
-        Append any existing Annotations to `self._annotations`.
-        Helper method for `add_annotation`, after a given Annotation
-        has been written to disk.
-        """
-        if self._memoize_annotations:
-            return self._memoize_annotations[0]
-
-        annotations = {}
-        annotations_dir = self._archiver.annotations_dir
-        # annotations_dir will be None for all previous archive versions < 7.0
-        if annotations_dir and os.path.exists(annotations_dir):
-            for annotation_id in os.listdir(annotations_dir):
-                annotation_path = os.path.join(annotations_dir, annotation_id)
-                annotation = Annotation.load(annotation_path)
-                annotations[annotation.name] = annotation
-
-        self._memoize_annotations.append(annotations)
-
-        return annotations
-
     def __init__(self):
         raise NotImplementedError(
             "%(classname)s constructor is private, use `%(classname)s.load`, "
@@ -191,7 +168,6 @@ class Result(IResult):
     def __new__(cls):
         result = object.__new__(cls)
         result._archiver = None
-        result._memoize_annotations = []
         return result
 
     def __repr__(self):
@@ -354,263 +330,23 @@ class Result(IResult):
                     yield (match.group(1), match.group(2))
 
     def add_annotation(self, annotation):
-        """
-        Add an Annotation onto a Result object.
-        All Result-associated parameters are passed into the sub-class's
-        `write` method, while the Annotation instance handles everything else
-
-        Parameters
-        ----------
-        annotation
-            An instantiated Annotation subclass (Note, Signature, etc).
-
-        Raises
-        ------
-        ValueError
-            If the Annotation name matches an existing Annotation name
-            attached to the Result in question.
-
-        Notes
-        -----
-            In Archive Format 7.0, `referenced_result_uuid` is set to
-            the same value as `root_result_uuid`, but this will change
-            in future versions to allow for Annotations that may reference
-            a different Result than the one they are attached to.
-
-        """
-        self._validate_annotation_support()
-        # Guard to ensure Annotation names are unique per Result object
-        if annotation.name in self._annotations:
-            raise ValueError(
-                'Duplicate name detected when attempting to add '
-                f'Annotation with name: "{annotation.name}"\n'
-                'Annotation names must be unique within each Result '
-                'they are attached to.'
-            )
-
-        annotation._write(annotations_dir=self._archiver.annotations_dir,
-                          root_result_uuid=str(self.uuid),
-                          referenced_result_uuid=str(self.uuid))
-        self._annotations[annotation.name] = annotation
-
-        # now calculate checksums for all files within the newly minted
-        # annotation subdir
-        # TODO: think about moving these into annotation._write after 7.1
-        annotation_dir = \
-            pathlib.Path(self._archiver.annotations_dir) / str(annotation.id)
-        checksum_ext = self._archiver._fmt.CHECKSUM_TYPE
-        manifest = self._archiver._fmt.CHECKSUM_FILE
-
-        checksums = util.checksum_directory(str(annotation_dir),
-                                            checksum_type=checksum_ext)
-
-        with (annotation_dir / manifest).open('w') as fh:
-            for item in checksums.items():
-                fh.write(util.to_checksum_format(*item))
-                fh.write('\n')
-
-        # this ensures additional attrs on Signature (signer name/email)
-        # are present when running Result.verify
-        loaded_annotation = Annotation.load(str(annotation_dir))
-        self._annotations[loaded_annotation.name] = loaded_annotation
+        self._archiver.add_annotation(annotation)
 
     def get_annotation(self, name):
-        """Retrieve an Annotation given by `name` from the Result object.
-
-        Parameters
-        ----------
-        name : str
-            The name of the Annotation to retrieve.
-
-        Returns
-        -------
-        Annotation : obj
-            The Annotation object associated with the provided name.
-
-        Raises
-        ------
-        KeyError
-            If no Annotation with the provided name is found.
-
-        """
-        self._validate_annotation_support()
-
-        if name in self._annotations:
-            return self._annotations[name]
-
-        raise KeyError(f'No Annotation with name: "{name}" was found.')
+        return self._archiver.get_annotation(name)
 
     def iter_annotations(self, filter_by_type=None):
-        """Constructs an iterable containing all Annotations associated with
-        the Result object.
-        """
-        self._validate_annotation_support()
-
-        if filter_by_type is None:
-            yield from self._annotations.values()
-        elif filter_by_type not in ANNOTATION_TYPE_DICT:
-            raise ValueError(f'Unknown annotation type: "{filter_by_type}". '
-                             'Supported annotation types are: '
-                             f'{ANNOTATION_TYPE_DICT.keys()}')
-        else:
-            for annotation in self._annotations.values():
-                if getattr(annotation, 'annotation_type') == filter_by_type:
-                    yield annotation
+        return self._archiver.iter_annotations(filter_by_type)
 
     def remove_annotation(self, name):
-        """
-        Remove an Annotation given by `name` from the Result object.
-
-        Parameters
-        ----------
-        name : str
-            The name of the Annotation to be removed.
-
-        Raises
-        ------
-        KeyError
-            If no Annotation with the specified name is found.
-
-        ValueError
-            If the corresponding annotation directory cannot be located.
-
-        """
-        self._validate_annotation_support()
-        annotations = self._annotations
-
-        # Guard against provided Annotation name not found on Result object
-        if name not in annotations:
-            raise KeyError(f'No Annotation found with name: "{name}"')
-
-        # Check for corresponding Annotation entry on disk
-        annotations_dir = self._archiver.annotations_dir
-        annotation_disk_dir = os.path.join(annotations_dir,
-                                           str(annotations[name].id))
-
-        if not os.path.exists(annotation_disk_dir):
-            raise ValueError('Unable to locate on-disk directory '
-                             f'for Annotation with name: "{name}"')
-
-        shutil.rmtree(annotation_disk_dir)
-        del annotations[name]
+        self._archiver.remove_annotation(name)
 
     def merge_annotations(self, other):
-        annotation_uuids = \
-            [str(annotation.id) for annotation in self._annotations.values()]
-
-        for other_annotation in other._annotations.values():
-            if str(other_annotation.id) not in annotation_uuids:
-                try:
-                    self.add_annotation(other_annotation)
-                except ValueError as e:
-                    if 'Duplicate name' in str(e):
-                        warnings.warn(f'Duplicate name {other_annotation.name}'
-                                      ' found. The annotation UUID will be'
-                                      ' prepended to the name of the new'
-                                      ' annotation.')
-                        # It should not be possible for this to collide because
-                        # we only get here if this annotation.id isn't present
-                        # on the artifact.
-                        other_annotation.name = \
-                            f'{other_annotation.name}-{other_annotation.id}'
-                        self.add_annotation(other_annotation)
-                    else:
-                        raise e
+        self._archiver.merge_annotations(other)
 
     def verify(self, signature_name):
-        """
-        Verify a Signature annotation by name on the provided Result.
-
-        Parameters
-        ----------
-        signature_name
-            Name of the Signature Annotation to verify.
-
-        Notes
-        -----
-        The following checks are performed:
-            - fingerprint match in local GPG keyring
-            - sha512sum for root level checksums file matches checksum_digest
-            - gpg detached signature verification
-            - sha512sum checks for each file in signature-level checksums file
-        """
-        # first make sure the Result isn't malformed & the Signature exists
         self.validate()
-        signature = self.get_annotation(signature_name)
-
-        annotation_dir = \
-            pathlib.Path(self._archiver.annotations_dir) / str(signature.id)
-
-        root_fp = self._archiver.root_dir
-        root_checksums_fp = root_fp / 'checksums.sha512'
-        sig_checksums_fp = annotation_dir / 'checksums.sha512'
-        signature_fp = annotation_dir / 'signature.gpg'
-
-        try:
-            fingerprint = getattr(signature, 'fingerprint')
-
-            if not fingerprint:
-                raise ValueError('Signature is missing fingerprint.')
-
-            found_fingerprint = gpg_find_key(fingerprint)
-            if not (normalize_fingerprint(found_fingerprint['fingerprint']) ==
-                    normalize_fingerprint(signature.fingerprint)):
-                raise ValueError('Found fingerprint does not match '
-                                 'fingerprint associated with Signature.')
-
-        except Exception as e:
-            raise ValueError(f'Signer key not found in local GPG keyring: {e}')
-
-        root_checksum_digest = sha512_file_hex(root_checksums_fp)
-        if not root_checksum_digest == getattr(signature, 'checksum_digest'):
-            raise ValueError(
-                'Root checksums.sha512 does not match digest in metadata.yaml')
-
-        try:
-            env = os.environ.copy()
-            unix_gpg_terminal_helper(env)
-
-            subprocess.run(
-                ['gpg', '--verify',
-                 str(signature_fp),
-                 str(root_checksums_fp)],
-                check=True, env=env,
-                stdout=subprocess.DEVNULL,
-                stderr=subprocess.PIPE,
-                text=True
-            )
-
-        except FileNotFoundError:
-            raise FileNotFoundError('`gpg` not found on PATH.')
-
-        except subprocess.CalledProcessError as e:
-            msg = (e.stderr or '').strip()
-            if len(msg) > 500:
-                msg = msg[:500] + '...'
-            raise subprocess.CalledProcessError(
-                f'`gpg --verify` failed (rc={e.returncode}): {msg}')
-
-        missing, mismatched = [], []
-        for exp_digest, relpath in self._iter_checksums(sig_checksums_fp):
-            fp = annotation_dir / relpath
-            if not fp.exists():
-                missing.append(relpath)
-                continue
-            obs_digest = sha512_file_hex(fp)
-            if obs_digest != exp_digest:
-                mismatched.append({
-                    'path': relpath,
-                    'expected': exp_digest,
-                    'actual': obs_digest
-                })
-        if missing:
-            raise ValueError('The following expected files were not found: '
-                             f'{missing}')
-        if mismatched:
-            raise ValueError('The following unexpected files were found: '
-                             f'{mismatched}')
-
-        return f'Signature `{signature_name}` verified successfully.'
+        return self._archiver.verify(signature_name)
 
 
 class Artifact(Result):
